refactor step3 OrderTableRepository -> table 도메인 패키지로 이동
